\name{coef}
\alias{coef}
\title{Dataset with coefficients}
\description{A dataset containing the coefficients fore computing the inverse Laplace transform in the Abate-Whitt frame work}
\usage{coef}
\format{
A list
\describe{\item{beta}{beta coefficients}
\item{eta}{eta coefficients}
}
}
\source{Telek}
