% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseLaplace1d.R
\name{inverseLaplace1d}
\alias{inverseLaplace1d}
\title{Inverse Laplace transform in 1D}
\usage{
inverseLaplace1d(t, h.star, ..., n = 15, n.cores = 1)
}
\arguments{
\item{t}{A vector of values for which to evaluate the inverse Laplace transform}

\item{h.star}{FUN Laplace transform to invert}

\item{...}{Additional arguments to h.star}

\item{n}{Accuracy parameter}

\item{n.cores}{Number of cores to use defaults to 1}
}
\value{
Inverse Laplace transform as a vector
}
\description{
Inverse Laplace transform in 1D
}
\examples{
inverseLaplace1d(1, function(s){s/(s^2+1)})
}
