% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mphDens.R
\name{mphDens}
\alias{mphDens}
\title{mphDens}
\usage{
mphDens(t1, t2, alpha, U, R, n.cores = 1, n = 15)
}
\arguments{
\item{t1}{First vector argument to inverse Laplace transform}

\item{t2}{Second vector argument to inverse Laplace transform}

\item{alpha}{Initial probability vector}

\item{U}{Inverse generator}

\item{R}{Reward matrix}

\item{n.cores}{Number of cores to be used}

\item{n}{Approximation accuary}
}
\value{
A vector of density evaluations in the points (t1,t2)
}
\description{
mphDens
}
\examples{
S <- matrix(c(-1.5, 0, 0,1.5, -1, 0,0, 1, -0.5), ncol = 3)
U <- solve(-S)
alpha <- c(0.2,0.4,0.4)
R <- matrix(c(0.2,0.5,0.8,0.4,0.1,0.9),nrow = 3)
mphDens(1, 1, alpha, U, R)
}
