% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mphDensity}
\alias{mphDensity}
\title{Inverse Laplace of 2D MPH Laplace transformed density}
\usage{
mphDensity(t1, t2, eta, beta, alpha, U, R)
}
\arguments{
\item{t1}{First argument in which to evaluate the density}

\item{t2}{Second argument in which to evaluate the density}

\item{eta}{Constants for Abate-Whitt}

\item{beta}{Constants for Abate-Whitt}

\item{alpha}{Inital probability vector}

\item{U}{Inverse of minus the generator matrix}

\item{R}{Reward matrix}
}
\value{
MPH density evaluation
}
\description{
Laplace inverse with Abate-Whitt frame work
}
